Updated CharacterCreation.cs, Updated GameWindow.cs, added StoryString.cs class
CharacterCreation now accesses the UI inside of the GameWindow.cs class. The MainWindowDialogue is now set to a different dialogue in CharacterCreation.

GameWindow was modified to not have a hardcoded string.

StoryString was added, to pass as a string inside of GameWindow. The intention is to be able to set new Srings inside of the UI, and then simply update the string when a story is needed.